Redirect user/placement to placement docs
The placement api-ref had a link to user/placement.html in nova. That is
being fixed in Ibfe016f25a29b6810ea09c5d03a01dbf3c53371f but the link
is now loose in the internets, using the 'latest' prefix, so we want to
make sure that what is currently showing as a 404 goes somewhere useful,
in this case the top of the placement docs.

Change-Id: Ib75547ac655e614441dec5e1bcc7784c2b6a070c