Remove "Resource Providers" upgrade check
With placement being extracted from nova, the
"Resource Providers" nova-status upgrade check no
longer works as intended since the placement data
will no longer be in the nova_api database. As a
result the check can fail on otherwise properly
deployed setups with placement extracted.

This check was originally intended to ease the upgrade
to Ocata when placement was required for nova to work,
as can be seen from the newton/ocata/pike references
in the code.

Note that one could argue the check itself as a concept
is still useful for fresh installs to make sure everything
is deployed correctly and nova-compute is properly
reporting into placement, but for it to be maintained
we would have to change it to no longer rely on the
nova_api database and instead use the placement REST API,
which while possible, might not be worth the effort or
maintenance cost.

For simplicity and expediency, the check is simply removed
in this change.

Related mailing list discussion can be found here [1].

[1] http://lists.openstack.org/pipermail/openstack-discuss/2018-December/000454.html

Change-Id: I630a518d449a64160c81410245a22ed3d985fb01
Closes-Bug: #1808956