Ensure rbd auth fallback uses matching credentials
As of Ocata, cinder config is preferred for rbd auth values with a
fallback to nova values [1]. The fallback path, for the case when
rbd_user is configured in cinder.conf and rbd_secret_uuid is not
configured in cinder.conf, results in the mismatched use of cinder
rbd_user with nova rbd_secret_uuid.

This fixes that fallback path to use nova rbd_user from nova.conf
with rbd_secret_uuid from nova.conf.

[1] See commit f2d27f6a8afb62815fb6a885bd4f8ae4ed287fd3

Thanks to David Ames for this fix.

Change-Id: Ieba216275c07ab16414065ee47e66915e9e9477d
Co-Authored-By: David Ames <david.ames@canonical.com>
Closes-Bug: #1809454