Add PrepResizeAtDestTask
This adds the sub-task to prep/verify the target host(s)
for the resize in the target cell. The PrepResizeAtDestTask
sub-task will make a synchronous RPC call (using
long_rpc_timeout) to method prep_snapshot_based_resize_at_dest
on the dest compute service which will claim resources on
the target host. The task also creates (inactive) port
bindings and volume attachments to be used on the target host.

If the prep task on the selected target host fails with a
MigrationPreCheckError, conductor will iterate over alternate
hosts and check them until a suitable target host is found
or we raise MaxRetriesExceeded.

The instance.migration_context is returned from the task so
it can be copied from the target DB to the source DB. This is
necessary for the API to route network-vif-plugged events later
when spawning the guest in the target cell.

Part of blueprint cross-cell-resize

Change-Id: I66d8f06f19c5c631e33208580428aa843abb38d2