Add PrepResizeAtSourceTask
This change adds a new conductor sub-task which will make a
synchronous RPC call (using long_rpc_timeout) to the new method
"prep_snapshot_based_resize_at_source" on the source compute.

If the instance is not volume-backed, the sub-task will create
an image and pass the image ID to the compute method to upload
the snapshot data.

If the migration fails at this point, any snapshot image created should
be deleted. Recovering the guest on the source host should be as simple
as hard rebooting the server (which is allowed with servers in ERROR
status).

Part of blueprint cross-cell-resize

Change-Id: I5bfcac018c1d1196d4efcb321213eb5a1d4c7a6b