Add TargetDBSetupTask
This adds a conductor task to create the instance related data
in the target cell prior to resizing (spawning) the instance in
the target cell. Since we need to preserve the history of the
instance in the target cell database, we have to explicitly copy
over things like instance actions and migration records.

Note that records created in the target cell database may have
a new id value so anything that needs to be accessed consistently
across both cells will need to rely on the UUID.

There are also some fields that we do not copy over since they
are specific to the source cell, like pci_devices and services.

Part of blueprint cross-cell-resize

Change-Id: I30916d8d10d70ce25523fa4961007cedbdfe8ad7