Reject resize with port having resource request
Nova does not consider the resource request of a Neutron port as of now.
So this patch makes sure that server resize request is rejected if it
involves a port that has resource request. When the feature is ready on
the nova side this limitation will be lifted.

Note that the api sample tests for server actions needed to split into
two test class. The floating ip tests are moved to a new class as those
tests are not supporting Neutron yet. For the rest neutron is now used
in the test. This was necessary as now the resize action depending on
neutron a neturon only feature, i.e. listing ports.

blueprint: bandwidth-resource-provider

Change-Id: I61a3e8902a891bac36911812e4e7c080570e3850