Reject migrate with port having resource request
Nova does not consider the resource request of a Neutron port as of now.
So this patch makes sure that server migrate and live migrate requests are
rejected if they involve a port that has resource request. When the feature
is ready on the nova side this limitation will be lifted.

blueprint: bandwidth-resource-provider

Change-Id: I48e6db9693e470b177bf4c75211d8b883c768433