Reject evacuate with port having resource request
Nova does not consider the resource request of a Neutron port as of now.
So this patch makes sure that server evacuate request is rejected if it
involves a port that has resource request. When the feature is ready on
the nova side this limitation will be lifted.

blueprint: bandwidth-resource-provider

Change-Id: Ic70d2bb781b6a844849a5cf2fe4d271b5a81093d