Reject unshelve with port having resource request
Nova does not consider the resource request of a Neutron port as of now.
So this patch makes sure that server unshelve request on servers that
was offload is rejected if it involves a port that has resource request.
When the feature is ready on the nova side this limitation will be lifted.

blueprint: bandwidth-resource-provider

Change-Id: I5a956513f3485074023e027430cc52ee7a3f92e4