Make functional-py37 job work like others
The functional-* jobs have specific dependencies (external
placement) which need to be present in the job. When the
functional-py37 was created it conflicted with that change
but since it is not yet run in the gate, we didn't notice.

Also, because the functional jobs use the tox-siblings
functionality in the gate, the tox name and tox env have to
be the same (that's just how tox-siblings works), so that
is fixed here too.

Change-Id: I51e60a3149d653753ec4fc9d62a8c670e6eaf96e