Fix incompatible version handling in BuildRequest
The BuildRequest object code assumed that IncompatibleObjectVersion
exception has a objver field that contains the object version. This
assumption is not true. The unit test made another mistake serializing
the function object obj_to_primitive instead of serializing the result
of the call of obj_to_primitive. This caused a false positive test
covering the error in the implementation as well.

Closes-Bug: #1812177

Change-Id: I1ef4a23aa2bf5cb46b481045f3d968f62f74606d