Add reshaper for PCPU
Added upgrade code using reshape to report PCPU instead of VCPU in case
the host is configured to use pinned CPUs. With this, when an existing
compute node running guests which uses dedicated CPUs is upgraded to
Train release, it will update allocation records of existing guest from
VCPU to PCPU using the reshape functionality.

Part of blueprint cpu-resources

Change-Id: I25d70aa09080b22d1bfa0aa097f0a114de8bf15a
Co-Authored-By: Stephen Finucane <sfinucan@redhat.com>