Add CrossCellMigrationTask
This adds the main conductor task that will orchestrate
a cross-cell resize. This will get called from the
existing conductor MigrationTask when that task determines
the scheduler picked a target host in another cell for the
resize operation.

The only thing CrossCellMigrationTask does in this change
is perform checks on the neutron and cinder APIs to make
sure they are new enough for what we need before we go any
further with the resize.

Part of blueprint cross-cell-resize

Change-Id: I4d181b44494f3b0b04537d5798537831c8fdf400