Fix using template cell urls with nova-manage
When nova-manage went to validate the transport-url given in config or on the
command line, it was not doing the translation before passing the url to the
oslo.messaging parse routine to check it. This exposes the format functions
from the CellMapping object, and makes our _validate_transport_url() format
the url before passing it to parse.

This also adds a test that makes sure the template makes it into the database
(as a template) and that it gets loaded out in translated form with an
object load.

Change-Id: I40a435b8e97c8552c2f5f0ca3a24de2edd9f81bd
Closes-Bug: #1812196
(cherry picked from commit edd1cd9ee4d149b01a388f03fb423374020033d7)