Allow utime call to fail on qcow2 image base file
This utime call sometimes fails, with EACCES (Permission Denied), when
the base file is on an NFS client filesystem.  I don't understand why,
but wonder if it's a similar problem as the one that motivated using
touch instead of utime in ec9d5e375e2.  In any case, IIUC, timing
isn't the primary thing that the image cache manager uses to determine
when the base file is in use.  The primary mechanism for that is
whether there is a matching disk file for a current instance.  The
timestamp on the base file is only used when deciding whether to
delete a base file that is _not_ in use; so it is not a big deal if
that deletion happens slightly earlier, for an unused base file,
because of one of these preceding utime calls having failed.

Closes-Bug: #1809123
Co-Authored-By: Matthew Booth <mbooth@redhat.com>
Change-Id: Idc131ff426f1707150867030fa5a69b77a7fc832