Cleanup return_reservation_id in ServersController.create
The create() method was getting the return_reservation_id
value out of the request body and putting it into the
create_kwargs dict just to pop it off, which is pointless.

This change simply avoids messing with the create_kwargs
dict for return_reservation_id and moves it to where it
is used so the comment block is removed since it is no
longer helpful.

Change-Id: I791ac7dc1630cebbac6abca9d31dc09b3787bc31