Add detection of SEV support from QEMU/AMD-SP/libvirt on AMD hosts
In order to support booting of SEV-enabled guests on AMD hosts which
are SEV-capable, we first need to be able to automatically detect
which hosts are capable of supporting SEV.  This requires checking for
support in two places:

- The sev parameter of the kvm_amd kernel module (which indicates that
  the kernel provides support)

- The host's libvirt domain capabilities, as indicated in the response
  from a virConnectGetDomainCapabilities() API call; if the <features>
  element includes <sev supported='yes'/> then this indicates that
  both QEMU and the AMD Secure Processor (AMD-SP) support SEV
  functionality, whereas <sev supported='no'/> indicates that QEMU
  doesn't support it, and the absence of any <sev> element indicates
  that libvirt doesn't even know about the feature:

    https://libvirt.org/git/?p=libvirt.git;a=commit;h=6688393c6b222b5d7cba238f21d55134611ede9c

This commit addresses the second of these two checks.  Subsequent
commits will address the first, add logic which combines both checks,
provide a new standard trait which indicates that both checks
have passed, and of course ultimately make use of that new trait.

blueprint: amd-sev-libvirt-support
Change-Id: I4c35b6a27db05349213429422ffe62adb09bd921