Move retry from _update to _update_to_placement
It was noted [1] that we don't need to wrap all of _update in a retry on
ResourceProviderUpdateConflict; only _update_to_placement can raise that
exception, and that exception only indicates a need to redrive with data
ret-GETted from within that method. So this patch moves the retry
decorator further in, to _update_to_placement, to avoid the additional
redundant compute node and pci tracker writes.

[1] https://review.openstack.org/#/c/615705/20/nova/compute/resource_tracker.py@974

Change-Id: I85e567181cdeefde0a9ef91cd460cd200e79bdba