doc: link admin/configuration from admin home page
There is some important stuff in the admin/configuration
docs sub-tree like information about configuring hypervisor
drivers and scheduler filters/weighers but it wasn't easily
found since it wasn't in the admin toc tree. This adds it
to the overall admin home page and adds a TODO that we need
to organize that admin page into sections somehow.

Change-Id: I5952a2dd590407b1ce56805df6f90a472cc878bf