Add prep_snapshot_based_resize_at_source compute method
This adds a new compute service method prep_snapshot_based_resize_at_source
which will be synchronously called over RPC from (super)conductor to the
source host during a cross-cell resize which will:

* Power off the instance
* Upload snapshot data if the instance is not volume-backed
* Delete the old BDM volume attachment records
* Destroy the guest on the hypervisor but retain disks in the
  case of a later fault or revert
* Activate the dest host port bindings

Think of this as a hybrid of how shelve_instance and resize_instance
work, with a lot of the flow matching what the migrate_disk_and_power_off
compute driver method does except for transferring disks to the dest host.

Also note that resources are not freed up from the ResourceTracker on
the source host at this point, nor are the instance.host/node values
nulled out since we are keeping a "placehoder" on the source in case
something fails later or the user reverts the resize.

Part of blueprint cross-cell-resize

Change-Id: I1887097ae38014dd19fb0ce333d7f223ad3d2130