LM: Use Claims to update numa-related XML on the source
With this patch, virt drivers get a new post_claim_migrate_data()
method. Currently, only the libvirt driver has a non-default
implementation. The method is named with future-proofing in mind, and
is meant to be called after a resources claim has been done on the
destination host of a live migration operation. Its purpose is to take
any information that the source host might need from the Claim, and
add it to migrate_data. For now, it is used to populate the
dst_numa_info field, which will tell the source how to update the XML
to match the resources (CPUs, hugepages, NUMA cells) that were claimed
on the destination.

On the source side of things, this patch adds a _update_numa_xml()
method. Given the current instance XML and the
LibvirtLiveMigrateNUMAInfo object from migrate_data.dst_num_info,
_update_numa_xml() updates things like CPU pins, NUMA cell pins,
and/or emulator thread pins to enable the instance to run correctly on
its destination host. _update_numa_xml() is not actually used by
anything yet, it will be wired up in subsequent patches.

Implements blueprint numa-aware-live-migration
Change-Id: I0f40f0461c34c49c6e26dcae5e9bd78ee7f1a720