Ignore some PendingDeprecationWarnings for os-vif
os-vif got bumped to 1.14.0 in https://review.openstack.org/572082 and
Nova is touching the VIFPortProfileOVSRepresentor class in a supported
fashion until at least Train.

However, that commit contributes to triggering bug 1813147 in check and
gate. The warning is safe to silence and should be converted to an error
in Train (once the older path is removed).

Change-Id: I9d8175c78f068cf9bdad4aeba0941f6ff49a9acc
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>
Related-Bug: #1813147
blueprint: generic-os-vif-offloads