Drop the integrated-gate (py27) template
We're currently running essentially redundant tempest
and grenade jobs on all changes in the check and gate
queues because we use the integrated-gate and
integrated-gate-py3 job templates which add the
following jobs:

* integrated-gate:
  - tempest-full
  - neutron-grenade

* integrated-gate-py3:
  - tempest-full-py3
  - grenade-py3

This change removes the integrated-gate template which
drops the tempest-full and neutron-grenade jobs from
nova changes. We will still have py27 coverage in unit
and functional test jobs which should be sufficient, and
both tempest and devstack still use the integrated-gate
template so they should detect any py27-specific breaking
changes.

Related ML thread:

http://lists.openstack.org/pipermail/openstack-discuss/2019-February/002388.html

Change-Id: I93e938277454a1fc203b3d930ec1bc1eceac0a1e