Add finish_snapshot_based_resize_at_dest compute method
This adds a new method to the nova-compute service which will
be synchronously RPC called from (super)conductor on the
destination host in the target cell during a cross-cell resize.

The method is similar to how the existing finish_resize method
works but parts are more like unshelve_instance, i.e. the use
of _prep_block_device, driver.spawn and deleting the snapshot
image (if there is one) after the guest is spawned.

Part of blueprint cross-cell-resize

Change-Id: I46728c191997da3c7b4c6ed9d6d3f9e6fb73ca3f