Don't set bandwidth limits for vhostuser, hostdev interfaces
'tc' [1] is used for rate limiting in libvirt. This is suitable for tap
devices but not for the sockets used by vhostuser interface or for
physical devices; rate limiting for these interfaces must be done in the
vSwitch or via the interface itself. For this reason, libvirt explicitly
marks these interfaces, along with some other interfaces types, as not
supporting bandwidth limiting [2]. Given that there is no reason to be
setting the bandwidth-related attributes on vhostuser or hostdev
interfaces, we should stop doing it.

[1] http://man7.org/linux/man-pages/man8/tc.8.html
[2] https://github.com/libvirt/libvirt/blob/568a41722/src/conf/netdev_bandwidth_conf.h#L38

Change-Id: Id57e4c149616fe0896e80fca322ece79ec3bfd69
Closes-Bug: #1814882