Fix InstanceMapping to always default queued_for_delete=False
This object has a default=False setting for queued_for_delete, but never
actually sets that value. All newly created records should have a non-NULL
value for this field, and we have a migration to fix them, so this
change explicitly forces that =False, unless the object is being created
with a value set.

Closes-Bug: #1814913
Change-Id: I99c5cc24c7e9bf5e2e72ffc868990b87b0e8e3f8