cleanup *.pyc files in docs tox envs
.pyc files are ignored by nova's .gitignore, as a result
if you change branches after the .pyc files are generated
for example from stable/ocata to master it does not remove
the .pyc files.

This can lead to the _validate_bytecode_header function that
is invoked as part of sphinx-build to fail as the .pyc
files no longer match the contents of the checked out files.

This change adds {[testenv]commands} to the docs envs
to clean up stale .pyc files before running sphinx-build

Change-Id: I536a260da639d32afbc998be220305f9489db375