Change live_migration_wait_for_vif_plug=True by default
This resolves the TODO to make the option default to True
so that the source compute service will wait for the
"network-vif-plugged" event, initiated by vif plugging during
pre_live_migration on the destination compute servie, before
initiating the guest transfer in the hypervisor. There are
certain networking backends that will not send the neutron
event for vif plugging alone (which is arguably a bug) but
OVS and linuxbridge, probably the two most widely used in
OpenStack deployments, are known to work with this config.

While in here, the Timeout message is fleshed out to give
more help with what the cause of the timeout could be and
possible recourse.

Change-Id: I8da38aec0fe4808273b8587ace3df9dbbc3ab576