Add FinishResizeAtDestTask
This adds the FinishResizeAtDestTask sub-task which
synchronously RPC calls finish_snapshot_based_resize_at_dest
on the destination compute service in the target cell
to finish the resize.

If that is successful, the "hidden" field on the source
and target cell instance record is swapped so that the API
shows the target cell instance and updates the instance
mapping record to point at the target cell since confirm/revert
actions at that point start in the target cell.

Part of blueprint cross-cell-resize

Change-Id: I3e28c0163dc14dacf847c5a69730ba2e29650370