Execute CrossCellMigrationTask from MigrationTask
This adds the code to check if a cross-cell move is
allowed by the RequestSpec, which is set in the API,
and if so, checks to see if the scheduler selected
a host in another cell. If so, then CrossCellMigrationTask
is executed.

Note the _restrict_request_spec_to_cell method gets
renamed and logging is adjusted since we may not be
restricting the scheduler to hosts in the same cell
if cross-cell moves are allowed.

Part of blueprint cross-cell-resize

Change-Id: Ibaa31fc8079c16e169472372c05dc48ff0602515